+++ src/tgtsnarf.c
@@ -245,7 +245,7 @@ upcase(char *string)
   char *p;
 
   for (p = string; *p != '\0'; p++)
-    *p = toupper(((unsigned char)(*p)));
+    *p = toupper((int)(unsigned char)*p);
 
   return (string);
 }
